create separate SaveDashboardButton for reuse, pass onSave
